+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix #1797: Utils.waitUntilReady should record the stack trace of the caller before rethrowing an exception
 * Add support for filtering labels by EXISTS/NOT_EXISTS via the single argument versions of `.withLabel` and `.withoutLabel`
 * Schedule reconnect in case of HTTP_GONE when watching; the rescheduled connect will start from beginning of history by not specifying resourceVersion
+* Example added for PortForward.
 
 #### Dependency Upgrade
 * Updated Knative Serving to v0.9.90